% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_auth.R
\name{gft_auth}
\alias{gft_auth}
\title{Authorize gft to access your Fusion Tables}
\usage{
gft_auth(cache = TRUE)
}
\arguments{
\item{cache}{One of logical or character. See \link[httr]{oauth2.0_token} for
more details}
}
\value{
Nothing if cache = TRUE
}
\description{
Authorize gft to access your Fusion Tables
}
