% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_create.R
\name{gft_create}
\alias{gft_create}
\title{Create a new Fusion Table from an R Object}
\usage{
gft_create(x, name, duplicate = FALSE, ...)
}
\arguments{
\item{x}{A dataframe or simple feature to be written to Google Fusion Table}

\item{name}{A string representing the name of the new fusion table}

\item{duplicate}{A logical, where TRUE allows Fusion Tables with the same
name (note that the ID is generated by Fusion Tables will be different even
if names match)}

\item{...}{Named arguments specifying path and file to direct
\link{gft_get_token} to the cached OAuth token or a token object}
}
\value{
A string of length 1, the Fusion Table ID
}
\description{
Upload an R object as a Fusion Table. If x is class \link[sf]{st_sf} then the
geometry will be converted to KML text prior to uploading for automatic
geo-referencing within Fusion Tables. Currently sf_to_kml() only supports
point features.
}
