% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_describe.R
\name{gft_describe}
\alias{gft_describe}
\title{Return Description Information For Fusion Table}
\usage{
gft_describe(table.id, ...)
}
\arguments{
\item{table.id}{A string represnting a Google Fustion Table ID}

\item{...}{Named arguments specifying path and file to direct
\link{gft_get_token} to the cached OAuth token or a token object}
}
\value{
A named list containing table and column information
}
\description{
Retreive desciptive information on a specific Fusion Table.
}
