% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_fetch.R
\name{gft_fetch}
\alias{gft_fetch}
\title{Return a tibble representing a Fusion Table}
\usage{
gft_fetch(table.id = NULL, query = NULL, token = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Return a tibble representing a Fusion Table
}
