% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_get_token.R
\name{gft_get_token}
\alias{gft_get_token}
\title{Fetch/Refresh a cached OAuth Token}
\usage{
gft_get_token(path = ".", file = ".httr-oauth", token = NULL)
}
\arguments{
\item{path}{A character string representing the path to a cached token}

\item{file}{A character string representing the filename of a cached token}

\item{token}{An object representing an OAuth token}
}
\description{
Fetch/Refresh a cached OAuth Token
}
