% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_list.R
\name{gft_list}
\alias{gft_list}
\title{List your Fusion Table IDs and Names}
\usage{
gft_list(...)
}
\arguments{
\item{...}{Named arguments specifying path and file to direct
\link{gft_get_token} to the cached OAuth token or a token object}
}
\value{
A dataframe with two columns \code{ID} and \code{name}
}
\description{
Return a dataframe containing the your Fusion Table IDs and names.
}
