% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_kml.R
\name{sf_to_kml}
\alias{sf_to_kml}
\alias{sf_to_kml.sfc_POINT}
\title{sf_to_kml}
\usage{
sf_to_kml(x)

\method{sf_to_kml}{sfc_POINT}(x)
}
\arguments{
\item{x}{A sf dataframe}
}
\value{
A dataframe with a new geometry column that is the KML text
representation of the sf geometry information
}
\description{
Convert sfc columns in an sf dataframe to KML text to allow automatic
georeferencing when uploading to Google Fusion Tables.
}
