% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_describe.R
\name{gft_describe}
\alias{gft_describe}
\title{Return Description Information For Fusion Table}
\usage{
gft_describe(table.id, token = NULL)
}
\arguments{
\item{token}{}
}
\description{
Return Description Information For Fusion Table
}
