% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_fetch.R
\name{gft_fetch}
\alias{gft_fetch}
\title{Return a tibble representing a Fusion Table}
\usage{
gft_fetch(table.id = NULL, query = NULL, ...)
}
\arguments{
\item{table.id}{A Fusion Table ID specifying which table to retreive data from}

\item{query}{A query to run on the selected table}

\item{...}{Named arguments specifying path and file to direct
\link{gft_get_token} to the cached OAuth token or a token object}
}
\value{
A dataframe of the query results
}
\description{
Run a SQL query on a fusion table and return the results as a dataframe
}
