% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_get_token.R
\name{gft_get_token}
\alias{gft_get_token}
\title{Fetch/Refresh a cached OAuth Token}
\usage{
gft_get_token(path = ".", file = ".httr-oauth")
}
\arguments{
\item{file}{}
}
\description{
Fetch/Refresh a cached OAuth Token
}
