% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_get_token.R
\name{gft_get_token}
\alias{gft_get_token}
\title{Fetch/Refresh a cached OAuth Token}
\usage{
gft_get_token(path = ".", file = ".httr-oauth",
  token = getOption("gft_token"))
}
\arguments{
\item{path}{A character string representing the path to a cached token}

\item{file}{A character string representing the filename of a cached token}

\item{token}{An object representing an OAuth token}
}
\value{
An object of class \link[httr]{Token}
}
\description{
If \code{token} is supplied then the same token is returned. If a path and
file are supplied then the token located there is loaded and refershed if
necessary.
}
