% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gft_update_col.R
\name{gft_update_col}
\alias{gft_update_col}
\title{Update a Fusion Table column's properties}
\usage{
gft_update_col(table.id, column, updates, token = NULL)
}
\arguments{
\item{token}{}
}
\description{
Update a Fusion Table column's properties
}
